+++ pdfg2d-font/src/main/java/net/zamasoft/pdfg2d/font/table/CoverageFormat1.java
@@ -43,4 +43,9 @@ public record CoverageFormat1(int[] glyphIds) implements Coverage {
 		}
 		return -1;
 	}
+
+	@Override
+	public int[] getGlyphIds() {
+		return this.glyphIds.clone();
+	}
 }